relocate enums, add lombok to resource owner
