+++ src/main/java/com/hw/clazz/eenum/ClientAuthorityEnum.java
@@ -1,4 +1,4 @@
-package com.hw.clazz;
+package com.hw.clazz.eenum;
 
 import com.hw.converter.DefaultAuthorityConverter;
